make the base controller have no ui, add subclass for one that has ui
